Not working, delete ClientThread.java, CustomThread.java
